# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (filtered by user's security groups)
      parameters:
        - name: parentId
          in: query
          schema:
            type: integer
            nullable: true
          description: Filter by parent task (null for root tasks)
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: assignedToId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: priorityId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: rootOnly
          in: query
          schema:
            type: boolean
          description: Return only root-level tasks (no parent)
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Task Notes
  /tasks/{taskId}/notes:
    get:
      operationId: listTaskNotes
      tags: [task-notes]
      summary: List notes for a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskNote"
    post:
      operationId: createTaskNote
      tags: [task-notes]
      summary: Add a note to a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskNoteRequest"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNote"
  /tasks/{taskId}/notes/{noteId}:
    put:
      operationId: updateTaskNote
      tags: [task-notes]
      summary: Update a note
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskNoteRequest"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNote"
    delete:
      operationId: deleteTaskNote
      tags: [task-notes]
      summary: Delete a note
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    SecurityGroupSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
...
    Task:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        priorityId:
          type: integer
          nullable: true
        priorityName:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        ownerName:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        progress:
          type: integer
          minimum: 0
          maximum: 100
        dueDate:
          type: string
          format: date
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        hasChildren:
          type: boolean
        childCount:
          type: integer
        noteCount:
          type: integer
        securityGroupIds:
          type: array
          items:
            type: integer
        securityGroups:
          type: array
          items:
            $ref: "#/components/schemas/SecurityGroupSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - progress
        - hasChildren
        - childCount
        - noteCount
        - securityGroupIds
        - securityGroups
        - createdAt
        - updatedAt
    CreateTaskRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        priorityId:
          type: integer
          nullable: true
        ownerId:
          type: integer
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        progress:
          type: integer
          minimum: 0
          maximum: 100
        dueDate:
          type: string
          format: date
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        securityGroupIds:
          type: array
          items:
            type: integer
      required:
        - name
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        priorityId:
          type: integer
          nullable: true
        ownerId:
          type: integer
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        progress:
          type: integer
          minimum: 0
          maximum: 100
        dueDate:
          type: string
          format: date
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        securityGroupIds:
          type: array
          items:
            type: integer
    TaskNote:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taskId
        - content
        - authorId
        - authorName
        - createdAt
        - updatedAt
    CreateTaskNoteRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...