# lib/api-zod/src/generated/types/taskNote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskNote {
  id: number;
  taskId: number;
  content: string;
  authorId: number;
  authorName: string;
  createdAt: Date;
  updatedAt: Date;
}